+++ kms-20160120/alibabacloud_kms20160120/models/_get_secret_value_request.py
@@ -9,41 +9,43 @@ class GetSecretValueRequest(DaraModel):
         self,
         dry_run: str = None,
         fetch_extended_config: bool = None,
+        recipient: str = None,
         secret_name: str = None,
         version_id: str = None,
         version_stage: str = None,
     ):
-        # Indicates whether to enable DryRun mode.
+        # Specifies whether to enable DryRun mode. Valid values:
         # 
-        # - true: Enabled  
-        # - false (Default Value): Disabled  
+        # - true: enables DryRun mode.
+        # - false (default): disables DryRun mode.
         # 
-        # DryRun mode is used for Testing API Calls to authenticate whether you have the required permissions on the specified resource and whether the Request Parameters are correctly configured. When DryRun mode is enabled, KMS always returns a failed response along with the failure reason. Possible failure reasons include:
+        # DryRun mode is used to test API calls and verify whether you have the required permissions on the corresponding resources and whether the request parameters are correctly configured. When DryRun mode is enabled, KMS always returns a failure and provides the failure reason. Failure reasons include:
         # 
-        # - DryRunOperationError: The request would succeed if the DryRun parameter were not specified.  
-        # - ValidationError: One or more parameters in the request are invalid.  
-        # - AccessDeniedError: You do not have permission to execute this operation on the KMS resource.
+        # - DryRunOperationError: The request would succeed without the DryRun parameter.
+        # - ValidationError: The parameters specified in the request are invalid.
+        # - AccessDeniedError: You are not authorized to perform this operation on the KMS resource.
         self.dry_run = dry_run
-        # Indicates whether to retrieve the extended configuration of the credential. Valid values:
+        # Specifies whether to retrieve the extended configuration of the secret. Valid values:
         # 
-        # - true: Retrieve  
-        # - false (Default Value): Do not retrieve  
+        # - true: retrieves the extended configuration.
+        # - false (default): does not retrieve the extended configuration.
         # 
-        # > Generic secrets do not support extended configuration. If you specify this parameter, it will be ignored.
+        # > Generic secrets do not support extended configurations. This parameter is ignored if specified.
         self.fetch_extended_config = fetch_extended_config
-        # The name or ARN of the credential.  
-        # > When accessing a credential under another Alibaba Cloud account, you must specify the credential ARN. The ARN format is `acs:kms:${region}:${account}:secret/${secret-name}`.
+        self.recipient = recipient
+        # The secret name or secret Alibaba Cloud Resource Name (ARN).
+        # >To access a secret in another Alibaba Cloud account, you must specify the secret ARN. The format of the secret ARN is `acs:kms:${region}:${account}:secret/${secret-name}`.
         # 
         # This parameter is required.
         self.secret_name = secret_name
-        # Version number.
+        # The version number.
         # 
-        # > The VersionId parameter is not supported for RDS credentials, PolarDB credentials, Redis/Tair credentials, RAM credentials, and ECS credentials. If you specify this parameter, it will be ignored.
+        # > ApsaraDB RDS secrets, PolarDB secrets, Redis/Tair secrets, RAM secrets, and ECS secrets do not support specifying VersionId. This parameter is ignored if specified.
         self.version_id = version_id
-        # The version stage. Default value: ACSCurrent.  
+        # The version stage. Default value: ACSCurrent.
         # 
-        # If you specify this parameter, the credential value of the specified version stage is returned. If you do not specify this parameter, the credential value of the ACSCurrent version stage is returned.  
-        # > For RDS credentials, PolarDB credentials, Redis/Tair credentials, RAM credentials, and ECS credentials, you can retrieve only the credential values corresponding to the ACSPrevious or ACSCurrent version stages.
+        # If you specify this parameter, the secret value of the specified version stage is returned. If you do not specify this parameter, the secret value of the ACSCurrent version stage is returned.
+        # > For ApsaraDB RDS secrets, PolarDB secrets, Redis/Tair secrets, RAM secrets, and ECS secrets, you can retrieve only the secret values of the ACSPrevious and ACSCurrent versions.
         self.version_stage = version_stage
 
     def validate(self):
@@ -59,6 +61,9 @@ class GetSecretValueRequest(DaraModel):
 
         if self.fetch_extended_config is not None:
             result['FetchExtendedConfig'] = self.fetch_extended_config
+
+        if self.recipient is not None:
+            result['Recipient'] = self.recipient
 
         if self.secret_name is not None:
             result['SecretName'] = self.secret_name
@@ -78,6 +83,9 @@ class GetSecretValueRequest(DaraModel):
 
         if m.get('FetchExtendedConfig') is not None:
             self.fetch_extended_config = m.get('FetchExtendedConfig')
+
+        if m.get('Recipient') is not None:
+            self.recipient = m.get('Recipient')
 
         if m.get('SecretName') is not None:
             self.secret_name = m.get('SecretName')